Refactor linking algorythm
Refactor linking algorythm as preparation to custom portal visiting order. We now do not link to ascending order of visited portals, but in order of longest distance to the current portal. This practically does not change in comparison to a default fan field, but enables correct fielding when walking back and forth. (hopefully)